Add item - tiled map - change level

+++ src/main/java/BaseObject.java
@@ -1,24 +1,24 @@
 import javafx.scene.shape.Rectangle;
 
 public abstract class BaseObject {
-    protected int x;
-    protected int y;
-    public BaseObject(int x, int y) {
+    protected double x;
+    protected double y;
+    public BaseObject(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x){
+    public void setX(double x){
         this.x = x;
     }
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public void setY(int y){
+    public void setY(double y){
         this.y = y;
     }
-    public int getY() {
+    public double getY() {
         return this.y;
     }
